requirements.txt
requests>=2.31.0
flask>=3.0.0
geopy>=2.4.1
matplotlib>=3.9.0
seaborn>=0.13.0
plotly>=5.18.0
pandas>=2.2.0
python-dotenv>=1.0.0
folium>=0.15.1
